pyproject.toml
[project]
name = "diplom"
version = "0.1.0"
description = ""
authors = [
    {name = "frozzzy1",email = "maleev.av@dclouds.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "environs (>=14.1.0,<15.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
